Fix test wibbles for new Template Haskell.
Because splices are now run in the renamer, we do not get the same error context
as we would when running in the type checker. In most cases we get less context,
and in some cases I have added additional context. Error messages should at
least tell the user that an error occurred in a splice; dropping context beyond
that point is not judged a great loss.

Note that we may now report only one error when multiple errors were reported
before because splices are now run in the renamer.